Fix event states updating - Re-add v1 parser
